+++ morph_validator/utils.py
@@ -2,27 +2,45 @@
 from pathlib import Path
 from typing import Dict, List
 
-from morph_tool.utils import neurondb_dataframe
+from morph_tool.utils import neurondb_dataframe, iter_morphology_files
 
 
-def get_valid_mtype_files(valid_mtype_db_file: Path,
-                          verify_path=True, ext='.h5') -> Dict[str, List[Path]]:
-    """Gets valid morphologies files.
+def get_mtype_files_db(
+        neurondb_file: Path, verify_path: bool = True, ext: str = '.h5') -> Dict[str, List[Path]]:
+    """Gets morphologies files defined by a neurondb file.
 
     Args:
-        valid_mtype_db_file: file of mappings between morphology name and mtype. Morphology files
+        neurondb_file: file of mappings between morphology name and mtype. Morphology files
         must be located in the same directory as this file.
+        verify_path: return only existing morphologies
+        ext: return morphologies with this extension only
 
     Returns:
-        dictionary of files per full mtype (mtype:msubtype)
+        dict: dictionary of files per full mtype (mtype:msubtype)
     """
-    df = neurondb_dataframe(valid_mtype_db_file)
-    df['path'] = df.apply(lambda x: Path(valid_mtype_db_file.parent, x['name'] + ext), axis=1)
+    df = neurondb_dataframe(neurondb_file)
+    df['path'] = df.apply(lambda x: Path(neurondb_file.parent, x['name'] + ext), axis=1)
     if verify_path:
         df = df[df.apply(lambda x: x['path'].exists(), axis=1)]
     mtype_dict = {mtype: list(set(df.path)) for mtype, df in df.groupby('mtype')}
 
     if not mtype_dict.keys():
-        raise ValueError('No mtypes in {}'.format(valid_mtype_db_file))
+        raise ValueError('No mtypes in {}'.format(neurondb_file))
 
     return mtype_dict
+
+
+def get_mtype_files_dir(dir_: Path) -> Dict[str, List[Path]]:
+    """Gets morphologies files defined by scanning a directory.
+
+    Args:
+        dir_: directory with morphologies. It must contain directories with morphologies files.
+            Those directories must be named after morphology type which files they contain.
+
+    Returns:
+        dict: dictionary of files per mtype
+    """
+    if not dir_.is_dir():
+        raise ValueError('"{}" must be a directory'.format(dir_))
+    return {mtype_dir.name: list(iter_morphology_files(mtype_dir))
+            for mtype_dir in dir_.iterdir()}